image badge & restructure
