+++ package.json
@@ -2,7 +2,6 @@
   "name": "reddit-poge",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://ghnew.github.io/reddit-poge/",
   "dependencies": {
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
