+++ src/components/Loader.jsx
@@ -1,0 +1,7 @@
+const Loader = () => (
+  <div className="loading">
+    <i className="fa fa-circle-o-notch fa-spin fa-fw fa-2x"></i>
+  </div>
+);
+
+export default Loader;
