+++ src/components/Preview.jsx
@@ -1,13 +1,17 @@
 import { useHotkeys } from "react-hotkeys-hook";
 
-const Preview = ({ img, handleClose, handleRight, handleLeft }) => {
+const Preview = ({ data : { video, image, type }, handleClose, handleRight, handleLeft }) => {
   useHotkeys("esc", handleClose, [handleClose]);
   useHotkeys("left", handleLeft, [handleLeft]);
   useHotkeys("right", handleRight, [handleRight]);
 	
   return (
     <div className="preview">
-      <img src={img} alt="" />
+      {type === 'video' ? (
+        <video controls autoPlay src={video}></video>
+      ) : (
+        <img src={image} alt="" />
+      )}
       <span onClick={handleClose} className="close">
         &#10005;
       </span>
