+++ src/constants.js
@@ -1,0 +1,4 @@
+export const Tab = {
+  Home: 0,
+  Fav: 1,
+};
