+++ src/styles/index.css
@@ -51,6 +51,7 @@ body {
 
 .grid-item {
   background-color: #eeee;
+  position: relative;
 }
 
 .grid-item img {
@@ -58,6 +59,27 @@ body {
   width: 100%;
   height: 200px;
   object-fit: contain;
+}
+
+.tag {
+  position: absolute;
+  right: 0;
+  background: #efdecd;
+  font-size: 12px;
+  font-weight: bold;
+  letter-spacing: 1px;
+}
+
+.tag.image {
+  background: #FF7F7F;
+}
+
+.tag.video {
+  background: #bae498;
+}
+
+.tag.gif {
+  background: #ADD8E6;
 }
 
 .sidebar {
